Conflicts:
	src/imports/dialogs/qquickmessagedialog.cpp
	src/imports/dialogs/qquickmessagedialog_p.h
	src/qml/debugger/qqmlprofilerservice_p.h
	src/qml/jsruntime/qv4regexpobject.cpp
	tests/auto/qml/debugger/qqmlprofilerservice/qqmlprofilerservice.pro

+++ src/qml/jsapi/qjsvalue.cpp
@@ -746,7 +746,7 @@ static bool js_equal(const QString &string, QV4::ValueRef value)
     if (value->isNumber())
         return __qmljs_string_to_number(string) == value->asDouble();
     if (value->isBoolean())
-        return __qmljs_string_to_number(string) == value->booleanValue();
+        return __qmljs_string_to_number(string) == double(value->booleanValue());
     if (value->isObject()) {
         Scope scope(value->objectValue()->engine());
         ScopedValue p(scope, __qmljs_to_primitive(value, PREFERREDTYPE_HINT));
